Refactor styles & types
